refactor(server): split oversized models module into SRP submodules
